feat(account-security): Split Security into separate items, make dedicated Security menu
